+++ src/main/java/com/shehab/udms/repo/NoticeRepo.java
@@ -3,7 +3,7 @@ package com.shehab.udms.repo;
 import com.shehab.udms.model.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeRepo extends JpaRepository<Notice,Integer> {
+public interface NoticeRepo extends JpaRepository<Notice,Long> {
 
 
 }
